add validations on service fields

+++ src/graphql/resolvers/serviceResolver.ts
@@ -19,10 +19,46 @@ import serviceRepository from '@repositories/serviceRepository';
 import { LoggedInContextType } from '@graphql/types';
 import userRepository from '@repositories/userRepository';
 import { createError } from '@errors/errorParser';
+import constants from '../../constants';
 
-const validateServiceModification = (
+const validateServiceArgs = (serviceAttributes: {
+    name?: string;
+    description?: string;
+    tags?: string[];
+    image_url?: string;
+}) => {
+    [
+        { attr: 'name', limit: 'maxNameLength' },
+        { attr: 'description', limit: 'maxDescriptionLength' },
+        { attr: 'tags', limit: 'maxTags' },
+        { attr: 'image_url', limit: 'maxImageUrlLength' },
+    ].forEach(({ attr, limit }) => {
+        const currentAttrLength = (serviceAttributes as any)[attr]?.length;
+        if (!currentAttrLength) return;
+
+        const attrMaxLength = (constants.serviceLimits as any)[limit];
+
+        if (currentAttrLength >= attrMaxLength) {
+            throw createError(
+                400,
+                `Service ${attr} length must be lower than ${attrMaxLength} (current ${currentAttrLength})`
+            );
+        }
+    });
+
+    serviceAttributes.tags?.forEach((tag) => {
+        if (tag.length > constants.serviceLimits.maxTagLength) {
+            throw createError(
+                400,
+                `Service tags lengths must be lower than ${constants.serviceLimits.maxTagLength} (current ${tag.length} on tag ${tag})`
+            );
+        }
+    });
+};
+
+const validateServiceModificationPermissions = (
     ctx: LoggedInContextType,
-    service: Service,
+    service: Service
 ) => {
     if (service.publisher_id !== ctx.userId && !ctx.isAdmin) {
         throw createError(403, "You don't have permissions for this action");
@@ -35,12 +71,14 @@ class ServiceResolver {
     @Mutation(() => Service)
     async createService(
         @Arg('creationArgs') creationArgs: CreateServiceArgs,
-        @Ctx() ctx: LoggedInContextType,
+        @Ctx() ctx: LoggedInContextType
     ): Promise<Service> {
+        validateServiceArgs(creationArgs);
+
         try {
             const res = await serviceRepository.addService(
                 creationArgs,
-                ctx.userId,
+                ctx.userId
             );
             return res;
         } catch (err: any) {
@@ -57,16 +95,15 @@ class ServiceResolver {
         @Ctx() ctx: LoggedInContextType,
         @Arg('after', { nullable: true }) paginationToken?: string,
         @Arg('query_term', { nullable: true }) queryTerm?: string,
-        @Arg('first', { nullable: true }) pageSize?: number,
+        @Arg('first', { nullable: true }) pageSize?: number
     ) {
-        const res = await serviceRepository.getServices(
+        return await serviceRepository.getServices(
             ctx.isAdmin,
             paginationToken,
             queryTerm,
             pageSize,
-            ctx.roles,
+            ctx.roles
         );
-        return res;
     }
 
     @Query(() => PaginatedServiceResponse)
@@ -75,13 +112,13 @@ class ServiceResolver {
         @Ctx() ctx: LoggedInContextType,
         @Arg('after', { nullable: true }) paginationToken?: string,
         @Arg('query_term', { nullable: true }) queryTerm?: string,
-        @Arg('first', { nullable: true }) pageSize?: number,
+        @Arg('first', { nullable: true }) pageSize?: number
     ) {
         const res = await serviceRepository.getServicesByPublisherId(
             ctx.userId,
             paginationToken,
             queryTerm,
-            pageSize,
+            pageSize
         );
         return res;
     }
@@ -91,23 +128,20 @@ class ServiceResolver {
     async updateService(
         @Arg('service_id') serviceId: string,
         @Arg('update_args') updateArgs: UpdateServiceArgs,
-        @Ctx() ctx: LoggedInContextType,
+        @Ctx() ctx: LoggedInContextType
     ) {
         const service = await serviceRepository.getServiceById(serviceId);
 
-        validateServiceModification(ctx, service);
+        validateServiceModificationPermissions(ctx, service);
+        validateServiceArgs(updateArgs);
 
-        const res = await serviceRepository.updateServiceById(
-            serviceId,
-            updateArgs,
-        );
-        return res;
+        return await serviceRepository.updateServiceById(serviceId, updateArgs);
     }
 
     @FieldResolver()
     async publisher(@Root() service: Service) {
         const publisher = await userRepository.getUserById(
-            service.publisher_id,
+            service.publisher_id
         );
         return publisher;
     }
@@ -116,14 +150,13 @@ class ServiceResolver {
     @Mutation(() => Service)
     async deleteService(
         @Arg('service_id') serviceId: string,
-        @Ctx() ctx: LoggedInContextType,
+        @Ctx() ctx: LoggedInContextType
     ): Promise<Service> {
         const service = await serviceRepository.getServiceById(serviceId);
 
-        validateServiceModification(ctx, service);
+        validateServiceModificationPermissions(ctx, service);
 
-        const res = await serviceRepository.deleteServiceById(serviceId);
-        return res;
+        return await serviceRepository.deleteServiceById(serviceId);
     }
 }
